#4 identifier valid or not
This function takes a string token and try to check all the rules for declaring identifiers.
R1. identifier should not be a keyword.
R2. first letter of the identifier should be in set of either a english letter or '_'(underscore).
R3. other characters can be in a set of english letters + underscore(_) + digits.